requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.5.2
attrs==22.1.0
black==24.10.0
click==8.1.7
Django==3.2
djangorestframework==3.12.4
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
idna==3.10
iniconfig==1.1.1
isort==5.13.2
line_profiler==4.1.3
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
Pillow==9.3.0
platformdirs==4.3.6
pluggy==0.13.1
py==1.11.0
pydantic==2.10.1
pydantic_core==2.27.1
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.6
sniffio==1.3.1
sqlparse==0.4.3
starlette==0.41.3
toml==0.10.2
tomli==2.1.0
typing_extensions==4.12.2
uvicorn==0.32.1
